Reorganise source
